add explicit range to holder_cutout_side
I found a small problem with "holder_cutout_side". Although it's documented as a percentage, it's clear from the code that it should be a decimal from 0.0 - 1.0. With the default of "0.0", OpenSCAD's UI will only let you enter integers. I was able to fix it in the SCAD file by adding an explicit range: